+++ src/main/java/net/servicestack/idea/AddServiceStackAction.java
@@ -25,22 +25,28 @@ import java.util.List;
 
 public class AddServiceStackAction extends AnAction {
 
-    public void actionPerformed(AnActionEvent e) {
+    public void actionPerformed(@NotNull AnActionEvent e) {
         Module module = getModule(e);
-        AddRef dialog = new AddRef(module);
+        AddRef dialog = new AddRef(module,e);
         dialog.pack();
         dialog.setLocationRelativeTo(null);
         dialog.setSize(dialog.getPreferredSize());
         dialog.setResizable(true);
         dialog.setTitle("Add ServiceStack Reference");
 
+        if(GradleBuildFileHelper.isUsingKotlin(e)) {
+            dialog.setDefaultNativeTypesHandler(new KotlinNativeTypesHandler());
+        } else if(GradleBuildFileHelper.isDartProject(module)) {
+            dialog.setDefaultNativeTypesHandler(new DartNativeTypesHandler());
+        } else {
+            dialog.setDefaultNativeTypesHandler(new JavaNativeTypesHandler());
+        }
+
         //Check if a package was selected in the left hand menu, populate package name
         PsiElement element = LangDataKeys.PSI_ELEMENT.getData(e.getDataContext());
         if (element instanceof PsiPackage) {
             PsiPackage psiPackage = (PsiPackage) element;
             dialog.setSelectedPackage(psiPackage);
-            dialog.setVisible(true);
-            return;
         }
 
         //Check if a directory containing a Java file was selected, populate package name
@@ -58,7 +64,7 @@ public class AddServiceStackAction extends AnAction {
                     PsiDirectory selectedDir = (PsiDirectory) element;
                     String packageName = "";
                     String moduleDirectoryPath = module.getModuleFile().getParent().getPath();
-                    List<String> packageArray = new ArrayList<String>();
+                    List<String> packageArray = new ArrayList<>();
                     while (selectedDir != null && !(moduleDirectoryPath.equals(selectedDir.getVirtualFile().getPath()))) {
                         packageArray.add(selectedDir.getName());
                         selectedDir = selectedDir.getParent();
@@ -76,7 +82,8 @@ public class AddServiceStackAction extends AnAction {
                     //do nothing, can't get package name.
                 }
             }
-            ShowDialog(module, dialog);
+
+            ShowDialog(module, dialog, e);
             return;
         }
 
@@ -102,14 +109,14 @@ public class AddServiceStackAction extends AnAction {
         //Check for document, display without a package name if no document.
         Document document = FileDocumentManager.getInstance().getDocument(selectedFile);
         if (document == null) {
-            ShowDialog(module, dialog);
+            ShowDialog(module, dialog,e);
             return;
         }
 
         //Check if a 'PsiFile', display without a package name if no PsiFile.
         PsiFile psiFile = PsiDocumentManager.getInstance(module.getProject()).getPsiFile(document);
         if (psiFile == null) {
-            ShowDialog(module, dialog);
+            ShowDialog(module, dialog,e);
             return;
         }
 
@@ -121,11 +128,11 @@ public class AddServiceStackAction extends AnAction {
                 dialog.setSelectedPackage(mainPackage);
             }
         }
-        ShowDialog(module, dialog);
+        ShowDialog(module, dialog,e);
     }
 
-    private void ShowDialog(Module module, AddRef dialog) {
-        if (GradleBuildFileHelper.isGradleModule(module) && GradleBuildFileHelper.isUsingKotlin(module)) {
+    private void ShowDialog(Module module, AddRef dialog, AnActionEvent event) {
+        if (GradleBuildFileHelper.isGradleModule(event) && GradleBuildFileHelper.isUsingKotlin(event)) {
             dialog.setFileName("dtos.kt");
         }
         else if (IDEAPomFileHelper.isMavenProjectWithKotlin(module)) {
@@ -165,7 +172,11 @@ public class AddServiceStackAction extends AnAction {
 
         boolean isMavenModule =  IDEAPomFileHelper.isMavenModule(module);
 
-        if (isAndroidProject(module) || isMavenModule || isDartProject(module)) {
+        if (isAndroidProject(module) ||
+                isMavenModule ||
+                GradleBuildFileHelper.isGradleModule(e) ||
+                isDartProject(module) ||
+                IsKotlinProject(e)) {
             e.getPresentation().setEnabled(true);
         } else {
             e.getPresentation().setEnabled(false);
@@ -197,6 +208,10 @@ public class AddServiceStackAction extends AnAction {
 
     private static boolean isDartProject(@NotNull Module module) {
         return GradleBuildFileHelper.isDartProject(module);
+    }
+
+    private static boolean IsKotlinProject(AnActionEvent event) {
+        return GradleBuildFileHelper.isUsingKotlin(event);
     }
 
     static Module getModule(AnActionEvent e) {