+++ src/main/java/net/servicestack/idea/common/AddTypeScriptAction.java
@@ -1,4 +1,4 @@
-package net.servicestack.idea;
+package net.servicestack.idea.common;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
@@ -10,8 +10,6 @@ import com.intellij.psi.PsiDirectory;
 import com.intellij.psi.PsiElement;
 import com.intellij.util.PlatformUtils;
 import org.jetbrains.annotations.NotNull;
-
-import java.io.File;
 
 /**
  * Created by Layoric on 28/05/2016.
@@ -27,33 +25,17 @@ public class AddTypeScriptAction extends AnAction {
         dialog.setResizable(true);
         dialog.setTitle("Add TypeScript ServiceStack Reference");
         PsiElement element = LangDataKeys.PSI_ELEMENT.getData(anActionEvent.getDataContext());
-        INativeTypesHandler defaultTsNativeTypesHandler = new TypeScriptNativeTypesHandler();
         if (element instanceof PsiDirectory) {
             PsiDirectory selectedDir = (PsiDirectory)element;
             dialog.setSelectedDirectory(selectedDir.getVirtualFile().getPath());
-            String initialName = getInitialFileName(selectedDir.getVirtualFile().getPath(),defaultTsNativeTypesHandler);
-            dialog.setInitialDtoName(initialName);
+            String initialName = "dtos";
+            dialog.setFileName(initialName);
         }
         showDialog(dialog);
     }
 
     private void showDialog(AddTypeScriptRef dialog) {
         dialog.setVisible(true);
-    }
-
-    private String getInitialFileName(String path, INativeTypesHandler defaultTsNativeTypesHandler) {
-        String initName = "ServiceReference";
-        int count = 1;
-        while(true) {
-            File existingFile = new File(path + "/" + initName + count +
-                    defaultTsNativeTypesHandler.getFileExtension());
-            if(existingFile.exists()) {
-                count++;
-            } else {
-                break;
-            }
-        }
-        return initName + count;
     }
 
     @Override