+++ src/main/java/net/servicestack/idea/common/NativeTypesLanguage.java
@@ -1,4 +1,4 @@
-package net.servicestack.idea;
+package net.servicestack.idea.common;
 
 /**
  * Created by Layoric on 4/12/2015.
@@ -9,5 +9,8 @@ public enum NativeTypesLanguage {
     Dart,
     TypeScriptConcrete,
     TypeScript,
-    Python
+    Python,
+    CSharp,
+    FSharp,
+    VbNet
 }