+++ CHANGELOG.md
@@ -5,8 +5,11 @@
 ## [Unreleased]
 ## [1.2.1-beta]
 - Add support for Python Add ServiceStack Reference in PyCharm.
-- Add support for TypeScript Add ServiceStack Reference in Rider.
 - Workaround plugin issue with 2021.1 IDEs and optional dependencies.
+- Higher res icons.
+- Fixes for Java Gradle projects.
+- Fixes for Kotlin project pathing.
+- Fixes for Android Gradle pathing.
 ## [1.2.0-nightly]
 - Add support for Python Add ServiceStack Reference in PyCharm.
 - Add support for TypeScript Add ServiceStack Reference in Rider.